requirements.txt
requests
openai
dotenv
csv
schedule
time
snowflake-connector-python